Set up Market class logic fully, so player can now buy and sell.
MapNode now chooses whether or not it is a market or event and GameManager toggles a menu depending on which the current node is.
Changed Market and Map class Start() functions to Awake()
Added functions to PlayerView for classes that need to edit player variables outside of the FSM and MVC.
Starting Player State is now Event instead of select.